% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalist.R
\name{datalistInput}
\alias{datalistInput}
\title{datalistInput}
\usage{
datalistInput(inputId, datalistId, label, options = c())
}
\arguments{
\item{inputId}{Input id}

\item{datalistId}{Datalist id}

\item{label}{Input label}

\item{options}{Vector of autocomplete options}
}
\description{
Creates an autocomplete input
}
